+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 
 ## 0.0.5(unreleased)
 1. fix `showTopTips` configure not restore to default
+2. fix `WeUIFormSelect ` not support `value` prop